Extract fake Temporal zoned date-time helper
